finish adding token location to parse tree
There are some more token locations that should probably be added
(for instance, the brackets in `var [a, b, c] = d;`), but I'm less
concerned about token locations now. They had been a crucial part of
my plan to associate comments with parsed structures, as is done in
golang (see [NewCommentMap](https://pkg.go.dev/go/ast@go1.21.0#NewCommentMap)).

However, my immediate goal is just to get comments to survive
transpilation, and I've come up with a radical approach: I'm going to
include comments in the syntax of the language, and only allow them
at specific locations.

This is not intended as a permanent solution; eventually comments
should be handled as in most languages (i.e., the parser should
ignore them).  But this temporary solution is a quick way to get
comments passed through to the output.